requirements.txt
absl-py==0.11.0
appdirs==1.4.4
APScheduler==3.6.3
astroid==2.4.2
astunparse==1.6.3
Brotli==1.0.9
cachetools==4.2.0
certifi==2020.12.5
chardet==3.0.4
chart-studio==1.1.0
click==7.1.2
colorama==0.4.3
colorhash==1.0.3
configparser==5.0.0
confuse==1.3.0
dash==1.18.1
dash-bootstrap-components==0.11.1
dash-core-components==1.14.1
dash-html-components==1.1.1
dash-renderer==1.8.3
dash-table==4.11.1
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
Flask-Compress==1.8.0
Flask-MonitoringDashboard==3.1.0
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.4.4
future==0.18.2
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.34.0
gunicorn==20.0.4
h5py==2.10.0
htmlmin==0.1.12
idna==2.10
ImageHash==4.1.0
imbalanced-learn==0.7.0
imblearn==0.0
importlib-metadata==3.1.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
joblib==1.0.0
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.4.3
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.2.2
mccabe==0.6.1
missingno==0.4.2
mysql-connector-python==8.0.22
mysqlclient==2.0.3
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.0.5
pandas-profiling==2.8.0
phik==0.10.0
plotly==4.14.1
protobuf==3.14.0
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.5.3
python-dateutil==2.8.1
pytz==2020.5
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
SQLAlchemy==1.3.22
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
toml==0.10.1
tqdm==4.47.0
typed-ast==1.4.1
tzlocal==2.1
urllib3==1.25.11
virtualenv==20.2.2
visions==0.4.4
webdriver-manager==3.2.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
